requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.0.3
joblib==1.5.3
numpy==2.4.2
packaging==26.0
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
whitenoise==6.11.0
